The better reading update
added an editor config file so other developers can follow the coding style. Moved CustomMessageBox into seperate controls folder. Added three fonts for later use. Added a social image. Updated control names in MainWindow and PlayerWindow to be easier to decern what they are and to. Added a text style selecter and a video width slider to MainWindow. Added #4 and #5.